+++ packages/generate-atomic-css/CHANGELOG.md
@@ -8,7 +8,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 ### Bug Fixes
 
 - Don't close the comment blocks above, it's not fully safe ([5ce6230](https://gitlab.com/codsen/codsen/commit/5ce6230))
-- Fix cases when there is no $$$ in the input and there are heads/tails ([09ceccd](https://gitlab.com/codsen/codsen/commit/09ceccd))
+- Fix cases when there is no \$\$\$ in the input and there are heads/tails ([09ceccd](https://gitlab.com/codsen/codsen/commit/09ceccd))
 - Fix the case when the only non-whitespace string is at zero index (falsey value) ([76fd296](https://gitlab.com/codsen/codsen/commit/76fd296))
 - More tests and more improvements to edge case recognition ([4721b20](https://gitlab.com/codsen/codsen/commit/4721b20))
 - Tightened the whitespace management part on the surrounding areas ([551a851](https://gitlab.com/codsen/codsen/commit/551a851))
@@ -21,7 +21,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 - Make more unit tests pass ([177f05c](https://gitlab.com/codsen/codsen/commit/177f05c))
 - opts.pad ([5d3e16a](https://gitlab.com/codsen/codsen/commit/5d3e16a))
 - Output the log ([c729689](https://gitlab.com/codsen/codsen/commit/c729689))
-- Tackle input without any $$$ ([1dcaee4](https://gitlab.com/codsen/codsen/commit/1dcaee4))
+- Tackle input without any \$\$\$ ([1dcaee4](https://gitlab.com/codsen/codsen/commit/1dcaee4))
 
 ## 1.0.0 (2018-06-28)